+++ universe/Pathfinder.h
@@ -1,0 +1,156 @@
+#ifndef _Pathfinder_h_
+#define _Pathfinder_h_
+
+
+#include "Enums.h"
+#include "ObjectMap.h"
+#include "TemporaryPtr.h"
+#include "UniverseObject.h"
+
+#include <boost/signals2/signal.hpp>
+#include <boost/unordered_map.hpp>
+#include <boost/numeric/ublas/symmetric.hpp>
+#include <boost/serialization/access.hpp>
+#include <boost/thread/shared_mutex.hpp>
+#include <boost/scoped_ptr.hpp>
+
+#include <vector>
+#include <list>
+#include <map>
+#include <string>
+#include <set>
+
+#ifdef FREEORION_WIN32
+// because the linker gets confused about Win32 API functions...
+#  undef GetObject
+#  undef GetObjectA
+#endif
+
+struct UniverseObjectVisitor;
+class XMLElement;
+class System;
+
+class Empire;
+struct UniverseObjectVisitor;
+class XMLElement;
+class ShipDesign;
+class System;
+
+/** The Pathfinder  class contains the locations of systems, the
+  * starlanes and functions to determine pathing and trip duration
+  * around the Universe. */
+class FO_COMMON_API Pathfinder {
+public:
+
+    /** \name Structors */ //@{
+    Pathfinder();                                     ///< default ctor
+    virtual ~Pathfinder();                            ///< dtor
+    //@}
+
+    /** \name Accessors */ //@{
+
+    /** Returns the straight-line distance between the objects with the given
+      * IDs. \throw std::out_of_range This function will throw if either object
+      * ID is out of range. */
+    double                  LinearDistance(int object1_id, int object2_id) const;
+
+    /** Returns the number of starlane jumps between the systems with the given
+      * IDs. If there is no path between the systems, -1 is returned.
+      * \throw std::out_of_range This function will throw if either system
+      * ID is not a valid system id. */
+    short                   JumpDistanceBetweenSystems(int system1_id, int system2_id) const;
+
+    /** Returns the number of starlane jumps between any two objects, accounting
+      * for cases where one or the other are fleets / ships on starlanes between
+      * systems. Returns INT_MAX when no path exists, or either object does not
+      * exist. */
+    //TODO smokeun int                     JumpDistanceBetweenObjects(int object1_id, int object2_id) const;
+
+    /** Returns the sequence of systems, including \a system1_id and
+      * \a system2_id, that defines the shortest path from \a system1 to
+      * \a system2, and the distance travelled to get there.  If no such path
+      * exists, the list will be empty.  Note that the path returned may be via
+      * one or more starlane, or may be "offroad".  The path is calculated
+      * using the visibility for empire \a empire_id, or without regard to
+      * visibility if \a empire_id == ALL_EMPIRES.
+      * \throw std::out_of_range This function will throw if either system ID
+      * is out of range, or if the empire ID is not known. */
+    std::pair<std::list<int>, double>
+                            ShortestPath(int system1_id, int system2_id, int empire_id = ALL_EMPIRES) const;
+
+    /** Returns the shortest starlane path distance between any two objects, accounting
+      * for cases where one or the other are fleets / ships on starlanes between
+      * systems. Returns -1 when no path exists, or either object does not
+      * exist. */
+    //TODO smokeun double                  ShortestPathDistance(int object1_id, int object2_id) const;
+
+    /** Returns the sequence of systems, including \a system1 and \a system2,
+      * that defines the path with the fewest jumps from \a system1 to
+      * \a system2, and the number of jumps to get there.  If no such path
+      * exists, the list will be empty.  The path is calculated using the
+      * visibility for empire \a empire_id, or without regard to visibility if
+      * \a empire_id == ALL_EMPIRES.  \throw std::out_of_range This function
+      * will throw if either system ID is out of range or if the empire ID is
+      * not known. */
+    std::pair<std::list<int>, int>
+                            LeastJumpsPath(int system1_id, int system2_id, int empire_id = ALL_EMPIRES,
+                                           int max_jumps = INT_MAX) const;
+
+    /** Returns whether there is a path known to empire \a empire_id between
+      * system \a system1 and system \a system2.  The path is calculated using
+      * the visibility for empire \a empire_id, or without regard to visibility
+      * if \a empire_id == ALL_EMPIRES.  \throw std::out_of_range This function
+      * will throw if either system ID is out of range or if the empire ID is
+      * not known. */
+    bool                    SystemsConnected(int system1_id, int system2_id, int empire_id = ALL_EMPIRES) const;
+
+    /** Returns true iff \a system is reachable from another system (i.e. it
+      * has at least one known starlane to it).   This does not guarantee that
+      * the system is reachable from any specific other system, as two separate
+      * groups of locally but not globally interconnected systems may exist.
+      * The starlanes considered depend on their visibility for empire
+      * \a empire_id, or without regard to visibility if
+      * \a empire_id == ALL_EMPIRES. */
+    bool                    SystemHasVisibleStarlanes(int system_id, int empire_id = ALL_EMPIRES) const;
+
+    /** Returns the systems that are one starlane hop away from system
+      * \a system.  The returned systems are indexed by distance from
+      * \a system.  The neighborhood is calculated using the visibility
+      * for empire \a empire_id, or without regard to visibility if
+      * \a empire_id == ALL_EMPIRES.
+      * \throw std::out_of_range This function will throw if the  system
+      * ID is out of range. */
+    std::multimap<double, int>              ImmediateNeighbors(int system_id, int empire_id = ALL_EMPIRES) const;
+
+    /** Returns the id of the System object that is closest to the specified
+      * (\a x, \a y) location on the map, by direct-line distance. */
+    int                                     NearestSystemTo(double x, double y) const;
+
+    //@}
+
+    /** \name Mutators */ //@{
+
+    /** Fills pathfinding data structure and determines least jumps distances
+      * between systems for the empire with id \a for_empire_id or uses the
+      * main / true / visible objects if \a for_empire_id is ALL_EMPIRES*/
+    void            InitializeSystemGraph(const std::vector<int> system_ids, int for_empire_id = ALL_EMPIRES);
+
+    /** Regenerates per-empire system view graphs by filtering the complete
+      * system graph based on empire visibility.  Does not regenerate the base
+      * graph to account for actual system-starlane connectivity changes. */
+    void            UpdateEmpireVisibilityFilteredSystemGraphs(int for_empire_id = ALL_EMPIRES);
+
+
+    //@}
+
+private:
+    class PathfinderImpl;
+    /* TODO make unique_ptr C++11 */
+    boost::scoped_ptr<PathfinderImpl> const pimpl;
+
+    /* friend class boost::serialization::access; */
+    /* template <class Archive> */
+    /* void serialize(Archive& ar, const unsigned int version); */
+};
+
+#endif // _Pathfinder_h_
